recipe-route-be/requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pathlib==1.0.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
pandas~=2.0.3
numpy~=1.25.1
scikit-learn~=1.3.0
fuzzywuzzy~=0.18.0
spacy~=3.6.0
nltk~=3.8.1
stripe~=5.5.0